合并homePage之前

+++ PS_downloadFile.py
@@ -3,7 +3,9 @@
 
 import requests
 import re
+import os
 from lxml import etree
+from requests.exceptions import ReadTimeout,ConnectionError,RequestException
 import rarfile
 
 import sys
@@ -11,6 +13,7 @@ reload(sys)
 sys.setdefaultencoding('gbk')
 #sys.setdefaultencoding('utf-8')
 
+import pyUnzip
 
 def getFileName(oURL):
     op = oURL.rfind('/')
@@ -25,6 +28,9 @@ def downLoadByURL(oUrl):
     html = requests.get(oUrl).text
     selector = etree.HTML(html)
     zipUrl  = selector.xpath('//ul[@class="downurllist"]/li/a/@href')
+    if not zipUrl:
+        return  False
+
     strZip = str(zipUrl[0])
     print strZip
 
@@ -35,36 +41,37 @@ def downLoadByURL(oUrl):
                 "User-Agent":"Mozilla/5.0 (Windows NT 6.1; WOW64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/42.0.2311.90 Safari/537.36"
                 }
 
-    req = requests.get(strZip,headers = oHeader)
-    print req
-
     localDir = 'D:\\pptSpider\\downloadFile\\'
     PDFName = getFileName(strZip)
     localPDF = localDir + PDFName
-    print localPDF
 
-    oFile = open(localPDF, "wb")
-    for chunk in req.iter_content(chunk_size=512):
-        if chunk:
-            oFile.write(chunk)
+    try:
+        response = requests.get(strZip,headers = oHeader ,timeout=10)
+        print(response.status_code)
+        print localPDF
 
-    unzipFile(PDFName)
+        #文件下载
+        if not os.path.exists(localPDF):
+            oFile = open(localPDF, "wb")
+            for chunk in response.iter_content(chunk_size=512):
+                if chunk:
+                    oFile.write(chunk)
+        else:
+            print "had download file:" ,localPDF
+
+        return True
+    except ReadTimeout:
+        print("timeout")
+    except ConnectionError:
+        print("connection Error")
+    except RequestException:
+        print("error")
 
 
-def unzipFile(fileName):
-    oFilePath = "D:\\pptSpider\\downloadFile\\" + fileName
-    fantasy_zip = rarfile.RarFile(oFilePath)
-    fantasy_zip.extractall("D:\\pptSpider\\downloadFile")
+    return False
+    #print type(PDFName)
+    #pyUnzip.unzipFile(PDFName)
 
-    '''
-    for oFileName in fantasy_zip.namelist():
-        filenameLower = oFileName.lower()
-        if filenameLower.find('.ppt') > 1:
-            print "find it:[" + oFileName + "]"
-            #fantasy_zip.extract(oFileName , "D:\\pptSpider\\downloadFile")
-            fantasy_zip.extractall(u"D:\\pptSpider\\downloadFile")
-    '''
-    fantasy_zip.close()
 
 def getContentByList(oListPageURL):
 
@@ -73,23 +80,23 @@ def getContentByList(oListPageURL):
 
     # listPage 的获取办法
     content3 = selector.xpath('//ul[@class="arclist"]/li/a/@href')
-    #for value in content3:
-    #    print value
+    for value in content3:
+        print value
 
     # 找到首页和尾页，得到所有内容.
-    content4 = selector.xpath(u'//ul[@class="pages"]/li/a')
+   # content4 = selector.xpath(u'//ul[@class="pages"]/li/a')
 
     #have nextpage 通过“下一页”的关键字查找
     #print u"//a[contains(text(), '下一页')]"
-    #content4 = selector.xpath(u"//a[contains(text(), '下一页')]")
+    content4 = selector.xpath(u"//a[contains(text(), '下一页')]")
     for value in content4:
         print value.text
 
     #decode("utf-8").encode('gbk')
 
-ppt1_url = 'http://www.1ppt.com/kejian/29032.html'
-downLoadByURL(ppt1_url)
-#unzipFile()
+#ppt1_url = 'http://www.1ppt.com/kejian/29032.html'
+#downLoadByURL(ppt1_url)
+
 
 #oUrl = 'http://www.1ppt.com/kejian/yuwen/138/kejian_1.html'
 #getContentByList(oUrl)
